% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extend0rnd1.R
\name{Extend0_n_rnd_groups}
\alias{Extend0_n_rnd_groups}
\alias{Extend0_with_n_rnd_groups}
\title{Add zero-frequency rows using complete and sampled group combinations}
\usage{
Extend0_n_rnd_groups(
  data,
  varGroups,
  n_rnd_groups = 0,
  rnd_rep = 1,
  rndSeed = 123
)

Extend0_with_n_rnd_groups(data, varGroups, non_rnd, rnd_rep = 1, rndSeed = 123)
}
\arguments{
\item{data}{data.frame within \code{\link{Extend0}}}

\item{varGroups}{argument to \code{\link{Extend0}}}

\item{n_rnd_groups}{The last \code{n_rnd_groups} elements of \code{varGroups} are
used as the basis for sampling.}

\item{rnd_rep}{When \code{rnd_rep = 1} (the default), the usual \code{Extend0}
extension is performed for the selected \code{varGroups} elements.
When \code{rnd_rep > 1}, this result is replicated to allow more combinations
to be sampled.}

\item{rndSeed}{Internal random seed to be used}

\item{non_rnd}{In \code{Extend0_with_n_rnd_groups}, the \code{varGroups} elements
that should not be sampled are specified by name. In this case,
\code{varGroups} must be a named list.}
}
\value{
A data frame.
}
\description{
\code{Extend0_with_n_rnd_groups} is a function that calls \code{Extend0()} with
\code{Extend0_n_rnd_groups()} as the \code{varGroups} attribute.
The data are extended in the usual way based on some of the \code{varGroups}
elements. For the remaining elements, combinations are sampled.
}
\examples{

# Data to be extended 
d <- SSBtoolsData("barcelona2025")[c(3, 6, 12, 18), -5]
rownames(d) <- NULL
d$year <- 2025:2026
d$month <- c("January", rep("August", 3))
d

# varGroups as a named list
varGroups <- list(geo = c("country", "city"), 
                  age = "age", 
                  sex = "sex", 
                  time = c("month", "year"))

a0 <- Extend0(d, varGroups = varGroups)
dim(a0)  # all combinations, 48 rows
 
a1 <- Extend0_with_n_rnd_groups(d, varGroups = varGroups, non_rnd = c("time", "geo"))
a1
dim(unique(a1[c("month", "year", "country", "city")])) # all combinations of time and geo  

unique(a1[c("sex", "month", "year")])        # not all combinations


a2 <- Extend0_with_n_rnd_groups(d, varGroups = varGroups, non_rnd = c("sex", "time", "geo"))
dim(a2)

# all combinations of selected variables 
dim(unique(a2[c("sex", "month", "year", "country", "city")]))
 
# not all combinations of selected variables 
dim(unique(a2[c("age", "month", "year", "country", "city")]))  
 

# effect of rnd_rep
for (rnd_rep in c(1, 2, 3, 5, 10, 50)) 
   print(dim(Extend0_with_n_rnd_groups(d, 
           varGroups = varGroups, non_rnd = c("time", "geo"), rnd_rep = rnd_rep)))
   
}
\seealso{
\code{\link[=Extend0rnd1]{Extend0rnd1()}}
}
