% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormulaSums.R
\name{FormulaSums}
\alias{FormulaSums}
\alias{Formula2ModelMatrix}
\title{Sums (aggregates) and/or sparse model matrix with possible cross table}
\usage{
FormulaSums(
  data,
  formula,
  makeNames = TRUE,
  crossTable = FALSE,
  total = "Total",
  printInc = FALSE,
  dropResponse = FALSE,
  makeModelMatrix = NULL,
  sep = "-",
  sepCross = ":",
  avoidHierarchical = FALSE,
  includeEmpty = FALSE,
  NAomit = TRUE,
  rowGroupsPackage = "base",
  viaSparseMatrix = TRUE,
  ...
)

Formula2ModelMatrix(data, formula, dropResponse = TRUE, ...)
}
\arguments{
\item{data}{data frame}

\item{formula}{A model formula}

\item{makeNames}{Column/row names made when TRUE}

\item{crossTable}{Cross table in output when TRUE}

\item{total}{Total code(s) which can be provided in several forms:
\itemize{
\item A single string (e.g. \code{"TOTAL"}) to be applied to all variables.
\item A \strong{named vector} or \strong{named list} giving one total code per variable,
e.g. \code{c(var1 = "ALL_A", var2 = "ALL_B")} or \code{list(var1 = "ALL_A", var2 = "ALL_B")}.
}}

\item{printInc}{Printing "..." to console when TRUE}

\item{dropResponse}{When TRUE response part of formula ignored.}

\item{makeModelMatrix}{Make model matrix when TRUE. NULL means automatic.}

\item{sep}{String to separate when creating column names}

\item{sepCross}{String to separate when creating column names involving crossing}

\item{avoidHierarchical}{Whether to avoid treating of hierarchical variables. Instead of logical, variables can be specified.}

\item{includeEmpty}{When \code{TRUE}, empty columns of the model matrix (only zeros) are included.
This is not implemented when a response term is included in the formula and \code{dropResponse = FALSE} (error will be produced).}

\item{NAomit}{When \code{TRUE}, NAs in the grouping variables are omitted in output and not included as a separate category.
Technically, this parameter is utilized through the function \code{\link{RowGroups}}.}

\item{rowGroupsPackage}{Parameter \code{pkg} to the function \code{\link{RowGroups}}.
Default is \code{"base"}.
Setting this parameter to \code{"data.table"} can improve speed.}

\item{viaSparseMatrix}{When TRUE, the model matrix is constructed by a single call to \code{\link[Matrix]{sparseMatrix}}.
Setting it to FALSE reverts to the previous behavior.
This parameter is included for testing purposes and will likely be removed in future versions.}

\item{...}{Further arguments to be passed to \code{FormulaSums}}
}
\value{
A matrix of sums, a sparse model matrix or a list of two or three elements (model matrix and cross table and sums when relevant).
}
\description{
By default this function return sums if the formula contains a response part and a model matrix otherwise
}
\details{
In the original version of the function the model matrix was constructed by
calling \code{\link[Matrix]{fac2sparse}} repeatedly.
Now this is replaced by a single call to \code{\link[Matrix]{sparseMatrix}}.
The sums are computed by calling \code{\link[stats]{aggregate}} repeatedly.
Hierarchical variables handled when constructing cross table.
Column names constructed from the cross table.
The returned model matrix includes the attribute \code{startCol} (see last example line).
}
\examples{
x <- SSBtoolsData("sprt_emp_withEU")

FormulaSums(x, ths_per ~ year*geo + year*eu)
FormulaSums(x, ~ year*age*eu)
FormulaSums(x, ths_per ~ year*age*geo + year*age*eu, crossTable = TRUE, makeModelMatrix = TRUE)
FormulaSums(x, ths_per ~ year:age:geo -1)
m <- Formula2ModelMatrix(x, ~ year*geo + year*eu)
print(m[1:3, ], col.names = TRUE)
attr(m, "startCol")
}
\seealso{
\code{\link{ModelMatrix}}
}
\author{
Øyvind Langsrud
}
