% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb_get_full_table.R
\name{ssb_get_full_table}
\alias{ssb_get_full_table}
\title{Fetch the full SSB table by requesting all combinations of values}
\usage{
ssb_get_full_table(table_id)
}
\arguments{
\item{table_id}{Table number, e.g. "12030"}
}
\value{
A data frame with all data from the table
}
\description{
Fetch the full SSB table by requesting all combinations of values
}
