% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbtools-package.R
\docType{package}
\name{ssbtools-package}
\alias{ssbtools}
\alias{ssbtools-package}
\title{ssbtools: Tools for accessing and caching data from Statistics Norway (SSB)}
\description{
This package simplifies downloading, caching, and working with data
from [SSB StatBank](https://www.ssb.no/en/statbank). You can save
your JSON queries from the SSB web UI and load them directly via
`ssb_get_csv()` or pull the entire dataset with `ssb_get_full_table()`.
}
\seealso{
\code{\link{ssb_get_csv}}, \code{\link{ssb_get_full_table}}, \code{\link{ssb_get_metadata}}
}
\author{
\strong{Maintainer}: Willem Ofstad \email{willem.ofstad@digdir.com}

}
\keyword{package}
