% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ssb_get_csv}
\alias{ssb_get_csv}
\title{Download or load SSB table as CSV with caching}
\usage{
ssb_get_csv(table_id, folder = "csv/ssb", json_folder = "JSON")
}
\arguments{
\item{table_id}{The table ID as string, e.g., "12030"}

\item{folder}{Where to save/read CSV files (default: "csv/ssb")}

\item{json_folder}{Folder containing query JSON files (default: "JSON")}
}
\value{
A data frame (tibble) from the API or CSV
}
\description{
Download or load SSB table as CSV with caching
}
