% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb_get_full_table.R
\name{ssb_get_full_table}
\alias{ssb_get_full_table}
\title{Fetch full table from SSB using table ID (all variables, all values)}
\usage{
ssb_get_full_table(table_id, language = "en")
}
\arguments{
\item{table_id}{The table ID as a string (e.g. "12030")}

\item{language}{"en" (English) or "no" (Norwegian). Default is "en"}
}
\value{
A data frame containing the full table (or as much as SSB allows)
}
\description{
Fetch full table from SSB using table ID (all variables, all values)
}
